WEB-79: (-) in progress converting 'help-menu-module' controllers
- modal works correctly, but 'show' for the separate window does not
