+++ custom/BU/js/outbound-link-logger.module.js
@@ -1,6 +1,6 @@
 let outboundLinksHelper = {
-  debug: false,
-  sendGAEvent: true,
+  debug: true,
+  sendGAEvent: false,
   logOutboundLinkMessage: function(message){
     if(this.debug){ console.log("outboundLinksLogger) " + message); }
   },
